some onboarding clean up
