+++ Sources/AudioPlayerClient/Live.swift
@@ -6,9 +6,11 @@ extension AudioPlayerClient {
       load: { sounds in
         .fireAndForget {
           queue.async {
+            let soundsToLoad = sounds.filter { !files.keys.contains($0) }
+
             try? AVAudioSession.sharedInstance().setCategory(.ambient)
             try? AVAudioSession.sharedInstance().setActive(true, options: [])
-            for sound in sounds {
+            for sound in soundsToLoad {
               for bundle in bundles {
                 guard let url = bundle.url(forResource: sound.name, withExtension: "mp3")
                 else { continue }
