+++ Sources/OnboardingFeature/OnboardingView.swift
@@ -238,9 +238,14 @@ public let onboardingReducer = Reducer<
 
   case .alert(.skipButtonTapped):
     state.alert = nil
-    return Effect(value: .alert(.confirmSkipButtonTapped))
-      .receive(on: ImmediateScheduler.shared.animation())
-      .eraseToEffect()
+    return .merge(
+      Effect(value: .alert(.confirmSkipButtonTapped))
+        .receive(on: ImmediateScheduler.shared.animation())
+        .eraseToEffect(),
+
+      environment.audioPlayer.play(.uiSfxTap)
+        .fireAndForget()
+    )
 
   case .delayedNextStep:
     state.step.next()
@@ -251,6 +256,10 @@ public let onboardingReducer = Reducer<
       environment.userDefaults
         .setHasShownFirstLaunchOnboarding(true)
         .fireAndForget(),
+
+      environment.audioPlayer.stop(.onboardingBgMusic)
+        .fireAndForget(),
+
       .cancel(id: DelayedNextStepId())
     )
 
@@ -346,6 +355,9 @@ public let onboardingReducer = Reducer<
     }
 
     return .merge(
+      environment.audioPlayer.load(AudioPlayerClient.Sound.allCases)
+        .fireAndForget(),
+      
       Effect
         .catching { try environment.dictionary.load(.en) }
         .subscribe(on: environment.backgroundQueue)
@@ -372,13 +384,19 @@ public let onboardingReducer = Reducer<
 
   case .nextButtonTapped:
     state.step.next()
-    return .none
+    return environment.audioPlayer.play(.uiSfxTap)
+      .fireAndForget()
 
   case .skipButtonTapped:
     guard !environment.userDefaults.hasShownFirstLaunchOnboarding else {
-      return Effect(value: .delegate(.getStarted))
-        .receive(on: ImmediateScheduler.shared.animation())
-        .eraseToEffect()
+      return .merge(
+        Effect(value: .delegate(.getStarted))
+          .receive(on: ImmediateScheduler.shared.animation())
+          .eraseToEffect(),
+        
+        environment.audioPlayer.play(.uiSfxTap)
+          .fireAndForget()
+      )
     }
     state.alert = .init(
       title: .init("Skip tutorial?"),
@@ -392,7 +410,8 @@ public let onboardingReducer = Reducer<
       secondaryButton: .default(.init("No, resume"), send: .resumeButtonTapped),
       onDismiss: .dismiss
     )
-    return .none
+    return environment.audioPlayer.play(.uiSfxTap)
+      .fireAndForget()
   }
 }
 .onChange(of: \.game.selectedWordString) { selectedWord, state, _, _ in
@@ -441,28 +460,6 @@ public let onboardingReducer = Reducer<
     return Effect(value: .delayedNextStep)
       .delay(for: 2, scheduler: environment.mainQueue.animation())
       .eraseToEffect()
-  }
-}
-.sounds()
-
-extension Reducer
-where State == OnboardingState, Action == OnboardingAction, Environment == OnboardingEnvironment {
-  func sounds() -> Self {
-    self.combined(
-      with: Self { _, action, environment in
-        switch action {
-        case .delegate(.getStarted):
-          return environment.audioPlayer.stop(.onboardingBgMusic)
-            .fireAndForget()
-
-        case .nextButtonTapped, .skipButtonTapped:
-          return environment.audioPlayer.play(.uiSfxTap)
-            .fireAndForget()
-
-        default:
-          return .none
-        }
-      })
   }
 }
 
@@ -571,69 +568,6 @@ private func isVisible(
     }
   }
 #endif
-
-extension GameState {
-  public static let onboarding = Self.init(
-    inProgressGame: .init(
-      cubes: .onboarding,
-      gameContext: .solo,
-      gameMode: .unlimited,
-      gameStartTime: Date(),
-      language: .en,
-      moves: [],
-      secondsPlayed: 0
-    )
-  )
-}
-
-extension Puzzle {
-  static var onboarding: Self {
-    var cubes = randomCubes(for: isowordsLetter).run()
-    cubes.1.2.2.left.letter = "G"
-    cubes.2.2.2.left.letter = "A"
-    cubes.2.2.2.right.letter = "M"
-    cubes.2.2.1.right.letter = "E"
-
-    cubes.1.2.2.top.letter = "C"
-    cubes.1.2.1.top.letter = "U"
-    cubes.2.2.2.top.letter = "B"
-    cubes.2.2.1.right.letter = "E"
-    cubes.2.2.1.top.letter = "S"
-
-    cubes.1.1.2.left.letter = "R"
-    cubes.2.1.2.left.letter = "E"
-    cubes.2.2.2.right.letter = "M"
-    cubes.2.1.2.right.letter = "O"
-    cubes.2.1.1.right.letter = "V"
-    cubes.2.2.1.right.letter = "E"
-
-    cubes.1.2.1.right.letter = "A"
-    cubes.2.1.1.top.letter = "M"
-    cubes.2.2.0.left.letter = "S"
-
-    cubes.0.2.0.top.letter = "P"
-    cubes.0.2.1.top.letter = "I"
-    cubes.0.2.2.top.letter = "L"
-    cubes.0.2.2.left.letter = "L"
-    cubes.0.1.2.left.letter = "O"
-    cubes.2.0.2.right.letter = "W"
-
-    cubes.0.0.2.left.letter = "W"
-    cubes.1.0.2.left.letter = "O"
-    cubes.2.0.2.left.letter = "R"
-    cubes.2.0.2.right.letter = "D"
-    cubes.2.0.1.right.letter = "S"
-
-    cubes.0.2.0.top.letter = "P"
-    cubes.1.2.0.top.letter = "U"
-    cubes.2.2.0.top.letter = "Z"
-    cubes.2.2.0.right.letter = "Z"
-    cubes.2.1.0.right.letter = "L"
-    cubes.2.0.0.right.letter = "E"
-
-    return cubes
-  }
-}
 
 extension String {
   var isRemove: Bool {
